+++ lib/main.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_siri_suggestions/flutter_siri_suggestions.dart';
 import 'package:lunasea/core.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 
@@ -59,7 +60,12 @@ class _State extends State<LunaBIOS> {
     @override
     void initState() {
         super.initState();
+        initSuggestions();
         SchedulerBinding.instance.addPostFrameCallback((_) => _boot());
+    }
+
+    void initSuggestions() async {
+        FlutterSiriSuggestions.instance.configure(onLaunch: (Map<String, dynamic> message) async {});
     }
 
     @override